requirements/default.txt
Django==3.1.7
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
django-rest-auth==0.9.5
factory-boy==3.2.0
ipython==7.21.0
psycopg2==2.8.6
python-decouple==3.4